refactor: 페어 리뷰 후에 리펙토링
- HomeController는 '/' 경로 forward만 해주고 회원가입폼, 게시글 작성 폼은 각자 컨트롤러로 수정
- Thymeleaf에서 fragment 경로 수정
- 자바 코드 Repository 네이밍 수정
