+++ src/main/java/kr/codesqaud/cafe/CafeApplication.java
@@ -1,4 +1,4 @@
-package kr.codesqaud;
+package kr.codesqaud.cafe;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -9,5 +9,4 @@ public class CafeApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(CafeApplication.class, args);
 	}
-
 }
