+++ src/main/java/kr/codesqaud/cafe/service/UserService.java
@@ -4,9 +4,9 @@ import kr.codesqaud.cafe.controller.dto.UserJoinDto;
 import kr.codesqaud.cafe.controller.dto.UserReadDto;
 import kr.codesqaud.cafe.controller.dto.UserUpdateDto;
 import kr.codesqaud.cafe.domain.User;
-import kr.codesqaud.cafe.exception.user.UserExceptionType;
-import kr.codesqaud.cafe.exception.user.UserJoinFailedException;
-import kr.codesqaud.cafe.exception.user.UserUpdateFailedException;
+import kr.codesqaud.cafe.common.exception.user.UserExceptionType;
+import kr.codesqaud.cafe.common.exception.user.UserJoinFailedException;
+import kr.codesqaud.cafe.common.exception.user.UserUpdateFailedException;
 import kr.codesqaud.cafe.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
