+++ src/main/resources/templates/home.html
@@ -1,8 +1,8 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
-<head th:replace="~{/fragment/layout/head :: common_head}"></head>
+<head th:replace="~{fragment/layout/head :: common_head}"></head>
 <body>
-<div th:replace="~{/fragment/layout/nav :: common_nav}"></div>
+<div th:replace="~{fragment/layout/nav :: common_nav}"></div>
 
 <div class="container" id="main">
    <div class="col-md-12 col-sm-12 col-lg-10 col-lg-offset-1">
